libia2: don't use `inline` fns (`thread_name_get`)
`inline` fns make it harder to reason about
what object file and compartment each function is from.
